fix: coverage; move types to *.d.ts files
