+++ .travis.yml
@@ -7,7 +7,10 @@ cache:
     - "node_modules"
 
 script:
+  - npm run lint
+  - npm run type-check
   - npm run test-with-coverage
+  - npm run build
 
 after_success:
   - npm run coverage
